source/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiokafka==0.12.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
catalogue==2.0.10
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
elasticsearch==7.13.0
elasticsearch7==7.17.9
exceptiongroup==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
jmespath==1.0.1
kafka-python==2.0.2
loguru==0.7.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
numpy==1.23.0
packaging==24.1
pandas==2.2.3
platformdirs==4.3.6
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==1.10.9
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-cache==1.2.1
rich==13.9.3
shellingham==1.5.4
six==1.16.0
SQLAlchemy==2.0.36
srsly==2.4.8
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.2
url-normalize==1.4.3
urllib3==1.26.20
win32-setctime==1.1.0
yarl==1.17.1
